overview of the rocket architecture; update patch
TODO:
The rocket overview is based on the current OpenRocket model.
Unfortunately, the OR model, the Onshape model, and the physical rocket
all disagree on certain dimensions. The models need to be scrubbed for
consistency! The fin can, nose cone, thrust flange, and recovery
micromodule are the main things I noticed.